While running the TestDeleteFile test the replication monitor was throwing errors like "trying to access storage ..."
the test was creating a file with under replicated blocks. when the file
was deleted it deleted all the corresponding rows from all the tables except
under_replicated_blocks table.

after some time replication monitor kicks in and it finds a row in the
under_replicated_Blocks table. every things starts to break down as
there are no corresponding rows in blocks_infos table etc etc

FIX was to delete the rows in under_replicated_blocks when deleting
the file.

NOTE optimized the delete file function. it was deleting inodes table
twice. there is still lot of room for optimization. Will fix it after
all the test are fixed